moved some variables
This commit removes the softsynth file descriptor and the debug flag
from the synth structure.  These will need to be accessed from a signal
handler and there is no way to pass the structure to it.
Also, the process_data function now works with local variables and assigns
them to the synth structure, avoiding allocating a buffer with malloc/free.