Fix spelling keystrokes and char-by-char echo.
Use ssml's interpret-as="characters" setting when the kernel reports
just one character. This allows the use of espeak's internationalized
spelling of letters instead of having to maintain spelling ourselves in speakup.

Original patch courtesy of
Samuel Thibault <samuel.thibault@ens-lyon.org>
and modified to work with the current code by Chris.

This fixes #6.