Add --volume option
This adds the command-line option to set the volume via --volume=N

When --volume is specified, the volume is locked using a new volumeSet flag to
prevent it from being overridden by runtime commands (e.g. from the kernel).

volumeMultiplier was also changed from 22 to 16, which maps the default volume to ~100%,
keeping the volume consistent when --volume isn't provided